% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP_EM_Algorithm.R
\name{DP_EM_algorithm}
\alias{DP_EM_algorithm}
\title{DP EM Algorithm Integrated Function}
\usage{
DP_EM_algorithm(Ymatrix, Xmatrix, t, N, iter, initial)
}
\arguments{
\item{Ymatrix}{matrix of dependent variables, multiple dimensions eligible}

\item{Xmatrix}{matrix of independent variables}

\item{t}{length of trajectory}

\item{N}{number of trajectories}

\item{iter}{number of maximum iterations}

\item{initial}{list of initialization values}
}
\value{
list of outputs used for M-step
}
\description{
Integration of E-step and M-step to iteratively update the parameters for DP
}
