% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP_EM_E_step.R
\name{DP_EM_E_step}
\alias{DP_EM_E_step}
\title{DP EM E-step Function}
\usage{
DP_EM_E_step(Ymatrix, Xmatrix, t, N, VarMatrix, Beta, K, Pcluster, VP, alpha)
}
\arguments{
\item{Ymatrix}{matrix of dependent variables, multiple dimensions eligible}

\item{Xmatrix}{matrix of independent variables}

\item{t}{length of trajectory}

\item{N}{number of trajectories}

\item{VarMatrix}{covariance matrix}

\item{Beta}{coefficient matrix}

\item{K}{number of clusters}

\item{Pcluster}{weights of clusters}

\item{VP}{sequence of i.i.d. random variables distributed as Beta(1,alpha), suppose to be infinite for DP, otherwise truncated by K}

\item{alpha}{hyperparameter with a constant value}
}
\value{
list of outputs used for M-step
}
\description{
E-step to calculate the expectation of log-likelihood in EM algorithm for DP
}
