% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Algorithm.R
\name{EM_algorithm}
\alias{EM_algorithm}
\title{EM Algorithm Integrated Function}
\usage{
EM_algorithm(Ymatrix, Xmatrix, split, t, N, p, np, iter, initial, lambda = 0)
}
\arguments{
\item{Ymatrix}{matrix of dependent variables, multiple dimensions eligible}

\item{Xmatrix}{matrix of independent variables}

\item{split}{index of columns with no penalty}

\item{t}{length of trajectory}

\item{N}{number of trajectories}

\item{p}{number of parameters}

\item{np}{number of groups of parameters}

\item{iter}{number of maximum iterations}

\item{initial}{list of initialization values}

\item{lambda}{penalty coefficient}
}
\value{
list of outputs used for M-step
}
\description{
Integration of E-step and M-step to iteratively update the parameters
}
