% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_E_step.R
\name{EM_E_step}
\alias{EM_E_step}
\title{EM E-step Function}
\usage{
EM_E_step(
  Ymatrix,
  Xmatrix,
  split,
  t,
  N,
  VarMatrix,
  Beta,
  K,
  Pcluster,
  Lambda = 0
)
}
\arguments{
\item{Ymatrix}{matrix of dependent variables, multiple dimensions eligible}

\item{Xmatrix}{matrix of independent variables}

\item{split}{index of columns with no penalty}

\item{t}{length of trajectory}

\item{N}{number of trajectories}

\item{VarMatrix}{covariance matrix}

\item{Beta}{coefficient matrix}

\item{K}{number of clusters}

\item{Pcluster}{weights of subgroups}

\item{Lambda}{penalty coefficient}
}
\value{
list of outputs used for M-step
}
\description{
E-step to calculate the expectation of log-likelihood in EM algorithm
}
