% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Data Simulation Function}
\usage{
data_sim(percentage, SDlist, t, Tmatrix, N, Coeflist, seed, random = TRUE)
}
\arguments{
\item{SDlist}{List of standard deviation matrices for each cluster}

\item{t}{Length of trajectory, t=1 will be single point;
t can be used to specify the length of trajectory for training/validation/testing}

\item{Tmatrix}{Feature matrix for the trajectory.}

\item{N}{Number of trajectories}

\item{Coeflist}{List of coefficient matrices for each cluster}

\item{seed}{Set the random seed for simulation}

\item{random}{Specify whether the simulation data has exact percentage
or approximate percentage for each cluster}

\item{percetage}{Percentage of each cluster, sum to 1}
}
\value{
Simulation dataset
}
\description{
Define the function to simulate the datasets with multiple dimensions and multivariate normal distributions.

percentage, SDlist, t, Tmatrix, N, Coeflist, seed=10, random = TRUE
}
\examples{
\donttest{
######single dimension simulation######
t <- c(1:8)
percentage <- c(0.7, 0.3)
N <- 200

# generate the feature matrix from polynomial formula
T <- c(1:8)
Tmatrix <- cbind(1, T, T^2, T^3)

# generate list of standard deviation matrices
SDlist = c(1, 1)

# generate list of coefficient matrices
# cluster 1
A1 <- c(100, 1, 1, 1)
# cluster 2
A2 <- c(80, 1, 1, 1)
Coeflist <- list(A1, A2)

# simulate the dataset
data_sim(percentage, SDlist, t, Tmatrix, N, Coeflist, seed=10, random = TRUE)

######multiple dimensions simulation######
t <- c(1:8)
percentage <- c(0.7, 0.3)
N <- 200

# generate the feature matrix from polynomial formula
T <- c(1:8)
Tmatrix <- cbind(1, T, T^2, T^3)

# generate list of standard deviation matrices
error_Sigma1 <- matrix(c(1,0,0,1),2,2)
error_Sigma2 <- matrix(c(1,0,0,1),2,2)
SDlist <- list(error_Sigma1,error_Sigma2)

# generate list of coefficient matrices
# cluster 1
A1 <- c(100,1,1,1)
B1 <- c(40,1,1,1)
# cluster 2
A2 <- c(80,1,1,1)
B2 <- c(30,1,1,1)
Coeflist <- list(cbind(A1, B1), cbind(A2, B2))

# simulate the dataset
data_sim(percentage, SDlist, t, Tmatrix, N, Coeflist, seed=10, random = TRUE)
}
}
